// frontend/lib/api/model/exerciseCreate.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Workout API
 * A workout tracking API built with FastAPI
 * OpenAPI spec version: 1.0.0
 */
import type { ExerciseModality } from "./exerciseModality";
import type { ExerciseCreatePictureUrl } from "./exerciseCreatePictureUrl";

/**
 * Schema for creating a new exercise.
 */
export interface ExerciseCreate {
  /**
   * Primary body part targeted
   * @minLength 1
   * @maxLength 100
   */
  body_part: string;
  /** Equipment/modality used */
  modality: ExerciseModality;
  /**
   * Exercise name
   * @minLength 1
   * @maxLength 255
   */
  name: string;
  /** URL to exercise demonstration image */
  picture_url?: ExerciseCreatePictureUrl;
}