// frontend/lib/api/model/exerciseModality.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Workout API
 * A workout tracking API built with FastAPI
 * OpenAPI spec version: 1.0.0
 */

/**
 * Enum for exercise modalities.
 */
export type ExerciseModality =
  (typeof ExerciseModality)[keyof typeof ExerciseModality];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ExerciseModality = {
  DUMBBELL: "DUMBBELL",
  BARBELL: "BARBELL",
  CABLE: "CABLE",
  MACHINE: "MACHINE",
  SMITH: "SMITH",
  BODYWEIGHT: "BODYWEIGHT",
} as const;